Allow coverage reporting to be visible to PRs
The problem is that the coverage report is being displayed on the command line, and it would be ideal to have a GitHub formatted output that shows up in code files
in the PR that highlights the lines of code that were not covered in the unit test CI workflow. Doing so, it would assist reviewers in ensuring that the code coverage is high.

The unit_tests workflow has been modified to:
- Create a coverage file only if the specified OS and python version finish the test
- Run a step to process the file and extract the file names and the missed lines/ranges and highlight the lines using workflow commands

Resolves: #288
Signed-off-by: youssef-itanii <youssef.itani@gmail.com>